move unneccessary code
