+++ src/core/entities/base_entity.py
@@ -2,7 +2,14 @@ from datetime import datetime
 
 from sqlmodel import Column, DateTime, Field, SQLModel, func
 
+
 class BaseEntity(SQLModel):
-    id: int = Field(primary_key=True)
-    created_at: datetime = Field(sa_column=Column(DateTime(timezone=True), default=func.now()))
-    updated_at: datetime = Field(sa_column=Column(DateTime(timezone=True), default=func.now(), onupdate=func.now()))
+    id: int = Field(default=None, nullable=False, primary_key=True)
+    created_at: datetime = Field(
+        sa_column=Column(DateTime(timezone=True), default=func.now())
+    )
+    updated_at: datetime = Field(
+        sa_column=Column(
+            DateTime(timezone=True), default=func.now(), onupdate=func.now()
+        )
+    )
