+++ src/core/entities/user.py
@@ -1,8 +1,8 @@
-from datetime import datetime
-from sqlalchemy import Column, DateTime, func
 from sqlmodel import Field, SQLModel
 
+from src.core.entities.base_entity import BaseEntity
 from src.models.auth.user_models import UserOut
+
 
 class UserBase(SQLModel):
     username: str = Field(default=None, nullable=False)
@@ -11,18 +11,8 @@ class UserBase(SQLModel):
     password_hash: str = Field(default=None, nullable=False)
 
 
-class User(UserBase, table=True):
+class User(BaseEntity, UserBase, table=True):
     __tablename__ = "users"
-
-    id: int = Field(default=None, nullable=False, primary_key=True)
-    created_at: datetime = Field(
-        sa_column=Column(DateTime(timezone=True), default=func.now())
-    )
-    updated_at: datetime = Field(
-        sa_column=Column(
-            DateTime(timezone=True), default=func.now(), onupdate=func.now()
-        )
-    )
 
 
 def user_entity_to_model(user: User) -> UserOut:
