+++ src/features/auth/fake_users_repository.py
@@ -1,5 +1,5 @@
 from src.core.contracts.users_repository_contract import UsersRepositoryContract
-from src.core.entities.auth.user import User, UserBase
+from src.core.entities.user import User, UserBase
 from sqlmodel.ext.asyncio.session import AsyncSession
