+++ src/features/auth/services/auth_service.py
@@ -9,7 +9,7 @@ from src.features.auth.utils.oauth2_utils import (
     verify_password,
 )
 from src.core.contracts.users_repository_contract import UsersRepositoryContract
-from src.core.entities.auth.user import UserBase, user_entity_to_model
+from src.core.entities.user import UserBase, user_entity_to_model
 from src.models.auth.user_models import Token, UserOut, UserRegister
