+++ src/main.py
@@ -1,15 +1,11 @@
 from fastapi import FastAPI
 from starlette.middleware.cors import CORSMiddleware
 
-from src.infrastructure.container import Container
-from src.routers.books import router as book_router
 from src.core.config import app_configs, settings
 from src.features.auth.auth_container import AuthContainer
 from src.features.auth.auth_routes import router as auth_router
 
 app = FastAPI(**app_configs)
-
-# container = Container()
 
 app.add_middleware(
     CORSMiddleware,
@@ -28,7 +24,6 @@ async def index() -> dict[str, str]:
 async def healthcheck() -> dict[str, str]:
     return {"status": "ok"}
 
-#app.include_router(book_router, tags=["Book"])
 
 auth_container = AuthContainer()
 app.include_router(auth_router, tags=["auth"])
